+++ src/subscriber/Subscriber.js
@@ -1,14 +1,13 @@
 import { MasterDB } from "@mov-ai/mov-fe-lib-core";
 
 class Subscriber {
-  constructor({ pattern, _onLoad, _onUpdate }) {
+  constructor({ pattern }) {
     this.pattern = pattern;
   }
 
   subscribe(onUpdate, onLoad) {
     MasterDB.subscribe(this.pattern, onUpdate, onLoad);
   }
-
   unsubscribe() {
     MasterDB.unsubscribe(this.pattern);
   }