+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # TBD
 
+- [FP-3488](https://movai.atlassian.net/browse/FP-3488): Unsubscribe to document basic data when document tab is closed
 - [FP-3050](https://movai.atlassian.net/browse/FP-3050): Toggle between XML and YAML isn't highlighting in real-time
 - [FP-3315](https://movai.atlassian.net/browse/FP-3315): IDE - New flow opens in the wrong split screen side